+++ src/event.c
@@ -239,7 +239,8 @@ static EVENT_CALLBACK(EVENT_HANDLER_APPLICATION_TERMINATED)
             struct view *view = window_manager_find_managed_window(&g_window_manager, window);
             if (view) {
                 space_manager_untile_window(&g_space_manager, view, window);
-                window_manager_remove_managed_window(&g_window_manager, window);
+                window_manager_remove_managed_window(&g_window_manager, window->id);
+                window_manager_purify_window(&g_window_manager, window);
             }
 
             if (g_mouse_state.window == window) g_mouse_state.window = NULL;
@@ -398,7 +399,8 @@ static EVENT_CALLBACK(EVENT_HANDLER_APPLICATION_HIDDEN)
         struct view *view = window_manager_find_managed_window(&g_window_manager, window);
         if (view) {
             space_manager_untile_window(&g_space_manager, view, window);
-            window_manager_remove_managed_window(&g_window_manager, window);
+            window_manager_remove_managed_window(&g_window_manager, window->id);
+            window_manager_purify_window(&g_window_manager, window);
         }
     }
 
@@ -481,7 +483,8 @@ static EVENT_CALLBACK(EVENT_HANDLER_WINDOW_DESTROYED)
     struct view *view = window_manager_find_managed_window(&g_window_manager, window);
     if (view) {
         space_manager_untile_window(&g_space_manager, view, window);
-        window_manager_remove_managed_window(&g_window_manager, window);
+        window_manager_remove_managed_window(&g_window_manager, window->id);
+        window_manager_purify_window(&g_window_manager, window);
     }
 
     return EVENT_SUCCESS;
@@ -612,7 +615,8 @@ static EVENT_CALLBACK(EVENT_HANDLER_WINDOW_MINIMIZED)
     struct view *view = window_manager_find_managed_window(&g_window_manager, window);
     if (view) {
         space_manager_untile_window(&g_space_manager, view, window);
-        window_manager_remove_managed_window(&g_window_manager, window);
+        window_manager_remove_managed_window(&g_window_manager, window->id);
+        window_manager_purify_window(&g_window_manager, window);
     }
 
     return EVENT_SUCCESS;